Move around and rename some tests
Test: changes tests
