+++ plugins/base/src/test/kotlin/translators/DefaultPsiToDocumentableTranslatorTest.kt
@@ -4,14 +4,12 @@ import org.jetbrains.dokka.base.DokkaBase
 import org.jetbrains.dokka.base.testApi.testRunner.BaseAbstractTest
 import org.jetbrains.dokka.links.DRI
 import org.jetbrains.dokka.model.Annotations
-import org.jetbrains.dokka.model.GenericTypeConstructor
 import org.jetbrains.dokka.model.TypeConstructor
 import org.jetbrains.dokka.model.doc.Text
 import org.jetbrains.dokka.model.firstMemberOfType
 import org.jetbrains.dokka.plugability.DokkaPlugin
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
-import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.Test
 import utils.assertNotNull
 
@@ -257,58 +255,6 @@ class DefaultPsiToDocumentableTranslatorTest : BaseAbstractTest() {
                     "String",
                     (kotlinSubclassFunction.parameters.firstOrNull()?.type as? TypeConstructor)?.dri?.classNames
                 )
-            }
-        }
-    }
-
-    @Disabled // The compiler throws away annotations on unresolved types upstream
-    @Test
-    fun `Can annotate UnresolvedBound`() {
-        testInline(
-            """
-            |/src/main/java/sample/FooLibrary.kt
-            |package sample;
-            |@MustBeDocumented
-            |@Target(AnnotationTarget.TYPE)
-            |annotation class Hello()
-            |fun bar(): @Hello() TypeThatDoesntResolve
-            """.trimMargin(),
-            configuration
-        ) {
-            documentablesMergingStage = { module ->
-                val type = module.packages.single().functions.single().type as GenericTypeConstructor
-                assertEquals(
-                    Annotations.Annotation(DRI("sample", "Hello"), emptyMap()),
-                    type.extra[Annotations]?.directAnnotations?.values?.single()?.single()
-                )
-            }
-        }
-    }
-
-    /**
-     * Kotlin Int becomes java int. Java int cannot be annotated in source, but Kotlin Int can be.
-     * This is paired with KotlinAsJavaPluginTest.`Java primitive annotations work`()
-     */
-    @Test
-    fun `Java primitive annotations work`() {
-        testInline(
-            """
-            |/src/main/java/sample/FooLibrary.kt
-            |package sample;
-            |@MustBeDocumented
-            |@Target(AnnotationTarget.TYPE)
-            |annotation class Hello()
-            |fun bar(): @Hello() Int
-            """.trimMargin(),
-            configuration
-        ) {
-            documentablesMergingStage = { module ->
-                val type = module.packages.single().functions.single().type as GenericTypeConstructor
-                assertEquals(
-                    Annotations.Annotation(DRI("sample", "Hello"), emptyMap()),
-                    type.extra[Annotations]?.directAnnotations?.values?.single()?.single()
-                )
-                assertEquals("kotlin/Int///PointingToDeclaration/", type.dri.toString())
             }
         }
     }
